Do not use int

+++ src/lib/oac-tree-gui-core/oac_tree_gui/transform/procedure_item_transform_helper.cpp
@@ -129,7 +129,7 @@ void PopulateProcedurePreambleItem(const preamble_t& preamble, ProcedurePreamble
 
   for (const auto& info : preamble.GetTypeRegistrations())
   {
-    item.AddTypeRegistration(static_cast<int>(info.GetRegistrationMode()), info.GetString());
+    item.AddTypeRegistration(info.GetRegistrationMode(), info.GetString());
   }
 }
